chore: update VIP number & simulate on recent block
